+++ packages/winnow/test/aggregate.js
@@ -5,25 +5,29 @@ const features = require('./fixtures/trees.json')
 const snowFeatures = require('./fixtures/snow.json')
 const budgetTable = require('./fixtures/budgetTable.json')
 
-test('Get a sum', (t) => {
+test('Get a sum', t => {
   t.plan(1)
   const options = {
-    aggregates: [{
-      type: 'sum',
-      field: 'Trunk_Diameter'
-    }]
+    aggregates: [
+      {
+        type: 'sum',
+        field: 'Trunk_Diameter'
+      }
+    ]
   }
   const results = winnow.query(features, options)
   t.equal(results.sum_Trunk_Diameter, 850305)
 })
 
-test('Use a group by', (t) => {
+test('Use a group by', t => {
   t.plan(3)
   const options = {
-    aggregates: [{
-      type: 'avg',
-      field: 'Trunk_Diameter'
-    }],
+    aggregates: [
+      {
+        type: 'avg',
+        field: 'Trunk_Diameter'
+      }
+    ],
     groupBy: 'Genus'
   }
   const results = winnow.query(features, options)
@@ -32,13 +36,15 @@ test('Use a group by', (t) => {
   t.ok(results[0].avg_Trunk_Diameter)
 })
 
-test('Use an order and a group by', (t) => {
+test('Use an order and a group by', t => {
   t.plan(2)
   const options = {
-    aggregates: [{
-      type: 'avg',
-      field: 'Trunk_Diameter'
-    }],
+    aggregates: [
+      {
+        type: 'avg',
+        field: 'Trunk_Diameter'
+      }
+    ],
     groupBy: 'Genus',
     order: 'avg_Trunk_Diameter DESC'
   }
@@ -47,7 +53,7 @@ test('Use an order and a group by', (t) => {
   t.equal(results[0].avg_Trunk_Diameter, 37)
 })
 
-test('Use a group by esri style', (t) => {
+test('Use a group by esri style', t => {
   t.plan(3)
   const options = {
     outStatistics: [
@@ -64,13 +70,15 @@ test('Use a group by esri style', (t) => {
   t.ok(results[0].avg_Trunk_Diameter)
 })
 
-test('Use multiple group bys', (t) => {
+test('Use multiple group bys', t => {
   t.plan(4)
   const options = {
-    aggregates: [{
-      type: 'avg',
-      field: 'Trunk_Diameter'
-    }],
+    aggregates: [
+      {
+        type: 'avg',
+        field: 'Trunk_Diameter'
+      }
+    ],
     groupBy: ['Genus', 'Common_Name']
   }
   const results = winnow.query(features, options)
@@ -80,71 +88,81 @@ test('Use multiple group bys', (t) => {
   t.ok(results[0].avg_Trunk_Diameter)
 })
 
-test('Get a named aggregate', (t) => {
+test('Get a named aggregate', t => {
   t.plan(1)
   const options = {
-    aggregates: [{
-      type: 'sum',
-      field: 'Trunk_Diameter',
-      name: 'total_diameter'
-    }]
+    aggregates: [
+      {
+        type: 'sum',
+        field: 'Trunk_Diameter',
+        name: 'total_diameter'
+      }
+    ]
   }
   const results = winnow.query(features, options)
   t.equal(results.total_diameter, 850305)
 })
 
-test('Get an aggregate on a field with a space', (t) => {
+test('Get an aggregate on a field with a space', t => {
   t.plan(1)
   const options = {
-    aggregates: [{
-      type: 'sum',
-      field: 'total precip'
-    }]
+    aggregates: [
+      {
+        type: 'sum',
+        field: 'total precip'
+      }
+    ]
   }
   const results = winnow.query(snowFeatures, options)
   t.equal(results.sum_total_precip, 135.69000000000003)
 })
 
-test('Get an aggregate on a field with a /', (t) => {
+test('Get an aggregate on a field with a /', t => {
   t.plan(1)
   const options = {
-    aggregates: [{
-      type: 'count',
-      field: 'Full/Part',
-      name: 'Full/Part_COUNT'
-    }]
+    aggregates: [
+      {
+        type: 'count',
+        field: 'Full/Part',
+        name: 'Full/Part_COUNT'
+      }
+    ]
   }
   const results = winnow.query(budgetTable, options)
   t.equal(results['Full/Part_COUNT'], 6885)
 })
 
-test('Get the variance of a field', (t) => {
+test('Get the variance of a field', t => {
   t.plan(1)
   const options = {
-    aggregates: [{
-      type: 'var',
-      field: 'total precip'
-    }]
+    aggregates: [
+      {
+        type: 'var',
+        field: 'total precip'
+      }
+    ]
   }
   const results = winnow.query(snowFeatures, options)
   t.equal(results.var_total_precip, 0.07661480700055341)
 })
 
-test('Get an aggregate with a where clause', (t) => {
+test('Get an aggregate with a where clause', t => {
   t.plan(1)
   const options = {
-    aggregates: [{
-      type: 'sum',
-      field: 'Trunk_Diameter',
-      name: 'total_diameter'
-    }],
+    aggregates: [
+      {
+        type: 'sum',
+        field: 'Trunk_Diameter',
+        name: 'total_diameter'
+      }
+    ],
     where: 'Trunk_Diameter > 10'
   }
   const results = winnow.query(features, options)
   t.equal(results.total_diameter, 735026)
 })
 
-test('Get multiple aggregates', (t) => {
+test('Get multiple aggregates', t => {
   t.plan(2)
   const options = {
     aggregates: [
@@ -165,7 +183,7 @@ test('Get multiple aggregates', (t) => {
   t.equal(results.max_diameter, 130)
 })
 
-test('Get multiple aggregates specified in the esri way', (t) => {
+test('Get multiple aggregates specified in the esri way', t => {
   t.plan(2)
   const options = {
     outStatistics: [
@@ -186,7 +204,7 @@ test('Get multiple aggregates specified in the esri way', (t) => {
   t.equal(results.max_diameter, 130)
 })
 
-test('Get multiple aggregates with a where clause', (t) => {
+test('Get multiple aggregates with a where clause', t => {
   t.plan(2)
   const options = {
     aggregates: [