+++ packages/winnow/test/fields.js
@@ -3,7 +3,7 @@ const test = require('tape')
 const winnow = require('../src')
 const features = require('./fixtures/trees.json').features
 
-test('Select a single field', (t) => {
+test('Select a single field', t => {
   t.plan(1)
   const options = {
     fields: ['Trunk_Diameter'],
@@ -14,7 +14,7 @@ test('Select a single field', (t) => {
   t.equal(fields[0], 'Trunk_Diameter')
 })
 
-test('Select a single field with string input', (t) => {
+test('Select a single field with string input', t => {
   t.plan(1)
   const options = {
     fields: 'Trunk_Diameter',
@@ -25,7 +25,7 @@ test('Select a single field with string input', (t) => {
   t.equal(fields[0], 'Trunk_Diameter')
 })
 
-test('Select multiple fields', (t) => {
+test('Select multiple fields', t => {
   t.plan(2)
   const options = {
     fields: ['Trunk_Diameter', 'Genus'],