+++ packages/winnow/test/project.js
@@ -114,9 +114,9 @@ test('Project to Web Mercator using 3857 with a geo filter', t => {
     limit: 1,
     geometry: {
       xmin: -20026376.39,
-      ymin: -20048966.10,
+      ymin: -20048966.1,
       xmax: 20026376.39,
-      ymax: 20048966.10,
+      ymax: 20048966.1,
       spatialReference: {
         wkid: 102100
       }
@@ -140,7 +140,8 @@ test('Project to Web Mercator using 102100', t => {
 test('Project to Web Mercator using WKT', t => {
   t.plan(1)
   const options = {
-    projection: 'PROJCS["WGS_1984_Web_Mercator_Auxiliary_Sphere",GEOGCS["GCS_WGS_1984",DATUM["D_WGS_1984",SPHEROID["WGS_1984",6378137.0,298.257223563]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.017453292519943295]],PROJECTION["Mercator_Auxiliary_Sphere"],PARAMETER["False_Easting",0.0],PARAMETER["False_Northing",0.0],PARAMETER["Central_Meridian",0.0],PARAMETER["Standard_Parallel_1",0.0],PARAMETER["Auxiliary_Sphere_Type",0.0],UNIT["Meter",1.0]]',
+    projection:
+      'PROJCS["WGS_1984_Web_Mercator_Auxiliary_Sphere",GEOGCS["GCS_WGS_1984",DATUM["D_WGS_1984",SPHEROID["WGS_1984",6378137.0,298.257223563]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.017453292519943295]],PROJECTION["Mercator_Auxiliary_Sphere"],PARAMETER["False_Easting",0.0],PARAMETER["False_Northing",0.0],PARAMETER["Central_Meridian",0.0],PARAMETER["Standard_Parallel_1",0.0],PARAMETER["Auxiliary_Sphere_Type",0.0],UNIT["Meter",1.0]]',
     limit: 1
   }
   const results = winnow.query(features, options)